+++ routes/author.php
@@ -17,8 +17,8 @@ Route::prefix('author')->name('author.')->group(function(){
         Route::view('/profile','back.pages.profile')->name('profile');
         Route::post('/change-profile-picture',[AuthorController::class, 'changeProfilePicture'])->name('change-profile-picture');
         Route::view('/settings','back.pages.settings')->name('settings');
-        Route::post('/change-blog-logo',[AuthorController::class, 'changeBlogLogo'])->name('change-blog-logo');
-        Route::post('/change-blog-favicon',[AuthorController::class, 'changeBlogFavicon'])->name('change-blog-favicon');
+        Route::post('/change-blog-logo',[AuthorController::class, 'changeBlogPic'])->name('change-blog-logo');
+        Route::post('/change-blog-favicon',[AuthorController::class, 'changeBlogPic'])->name('change-blog-favicon');
     });
 
 });